+++ src/Control/Lens/Getter.hs
@@ -47,6 +47,7 @@ module Control.Lens.Getter
   , to
   -- * Combinators for Getters and Folds
   , (^.), (^$)
+  , (%), (^%)
   , view
   , views
   , use
@@ -65,8 +66,35 @@ import Control.Monad.State.Class        as State
 -- $setup
 -- >>> import Control.Lens
 
-infixl 8 ^.
+infixl 8 ^., ^%
+infixl 1 %
 infixr 0 ^$
+
+-------------------------------------------------------------------------------
+-- Pipelining
+-------------------------------------------------------------------------------
+
+-- | Passes the result of the left side to the function on the right side (forward pipe operator).
+--
+-- This is the flipped version of ('$'), which is more common in languages like F# as (@|>@) where it is needed
+-- for inference. Here it is supplied for notational convenience and given a precedence that allows it
+-- to be nested inside uses of ('$').
+--
+-- >>> "hello" % length % succ
+-- 6
+(%) :: a -> (a -> b) -> b
+a % f = f a
+{-# INLINE (%) #-}
+
+-- | A version of ('Control.Lens.Combinators.%') with much tighter precedence that can be interleaved with ('^.')
+--
+-- >>> "hello"^%length
+-- 5
+-- >>> import Data.List.Lens
+-- >>> ("hello","world")^._1^%reverse^._head
+-- 'o'
+(^%) :: a -> (a -> b) -> b
+a ^% f = f a
 
 -------------------------------------------------------------------------------
 -- Getters
@@ -101,6 +129,7 @@ type Getter a c = forall f. Gettable f => (c -> f c) -> a -> f a
 to :: (a -> c) -> Getter a c
 to f g = coerce . g . f
 {-# INLINE to #-}
+
 
 -- |
 -- Most 'Getter' combinators are able to be used with both a 'Getter' or a