+++ CHANGELOG.markdown
@@ -1,7 +1,8 @@
 2.9
 -----
 * Added `<<%~`, `<<.~`, `<<%=` and `<<.=` for accessing the old values targeted by a `Lens` (or a summary of those targeted by a `Traversal`)
-* Renamed `|>` to `%`, as `%~` is the lensed version of `%`.
+* Renamed `|>` to `%`, as `%~` is the lensed version of `%`, and moved it to `Control.Lens.Getter` along with a version `^%` with tighter
+  precedence that can be interleaved with `^.`
 * Upgraded to `doctest` 0.9, which lets us factor out common `$setup` for our doctests
 
 2.8